pipeline-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dateutil==2.8.2
pydantic[email]==2.4.2
